feat(web): split methodology into reading tab and consolidated evidence tab
